# GH-72904: Simplify implementation of `fnmatch.translate()`