Use `re` to scan shell-style patterns, rather than parsing them by hand in a fat loop. This makes the code twice as slow (!) but arguably more obvious


<!-- gh-issue-number: gh-72904 -->
* Issue: gh-72904
<!-- /gh-issue-number -->