@AA-Turner 2023-09-27
> twice as slow (!)

What are the benchmarks for this? (i.e. are we talking 10μs to 20μs or 250ms to 500ms?) The regression in performance here may not matter if the translation is a sufficiently small part of any reasonable usecase.

@serhiy-storchaka has been working on ``re`` performance lately and may have tips?

A