@barneygale 2023-10-29
Withdrawing this PR - the performance isn't good enough.